Fix 14 reliability issues and add 36 new tests
Millis rollover safety: replace absolute-target timers with elapsed-time
pattern across .ino, APIUploader, and WebServer. Thread safety: add
spinlock for shared timing globals and I2C mutex for dual-core access.
Flash wear: batch metadata saves every 50 writes, crash-safe trim with
backup-first rename. Fix uint16_t overflow in PumpController, GPS NaN
guards in payload and sensor records, DO compensation ordering, 30-min
continuous mode auto-timeout, pump config wiring at startup, PROGMEM
for HTML pages, and web endpoint pump config persistence.

+++ SeaSenseLogger/src/pump/PumpController.h
@@ -44,9 +44,9 @@ struct PumpConfig {
     uint8_t measurementCount;         // 1 default
     uint16_t measurementIntervalMs;   // 2000ms default
     uint16_t pumpStopDelayMs;         // 500ms default
-    uint16_t cooldownMs;              // 55000ms default
+    uint32_t cooldownMs;              // 55000ms default
     unsigned long cycleIntervalMs;    // 60000ms (1 minute) default
-    uint16_t maxPumpOnTimeMs;         // 30000ms safety
+    uint32_t maxPumpOnTimeMs;         // 30000ms safety
     uint8_t relayPin;                 // GPIO 25
     bool enabled;                     // true default
